feat: DTO와 예외 클래스를 별도의 클래스 파일로 분리
