+++ replace/src/main/kotlin/com/app/replace/application/response/ImageURLRecord.kt
@@ -1,0 +1,5 @@
+package com.app.replace.application.response
+
+import com.fasterxml.jackson.annotation.JsonValue
+
+data class ImageURLRecord(@JsonValue val imageURL: String)
