+++ replace/src/main/kotlin/com/app/replace/application/response/SimpleUserInformation.kt
@@ -1,0 +1,3 @@
+package com.app.replace.application.response
+
+data class SimpleUserInformation(val nickname: String, val imageUrl: String)
