+++ replace/src/main/kotlin/com/app/replace/application/response/SimpleUserProfile.kt
@@ -1,0 +1,3 @@
+package com.app.replace.application.response
+
+data class SimpleUserProfile(val nickname: String, val profileImage: String)
