+++ replace/src/main/kotlin/com/app/replace/application/response/Writer.kt
@@ -1,0 +1,3 @@
+package com.app.replace.application.response
+
+data class Writer(val profileImage: String, val nickname: String)
