+++ replace/src/main/kotlin/com/app/replace/ui/request/CreateDiaryRequest.kt
@@ -1,0 +1,3 @@
+package com.app.replace.ui.request
+
+data class CreateDiaryRequest(val title: String, val content: String, val shareScope: String, val images: List<String>)
