+++ replace/src/main/kotlin/com/app/replace/ui/request/ImageUploadingRequest.kt
@@ -1,0 +1,5 @@
+package com.app.replace.ui.request
+
+import org.springframework.web.multipart.MultipartFile
+
+data class ImageUploadingRequest(val images: List<MultipartFile>)
