+++ replace/src/main/kotlin/com/app/replace/ui/request/MakingConnectionRequest.kt
@@ -1,0 +1,3 @@
+package com.app.replace.ui.request
+
+data class MakingConnectionRequest(val code: String)
