+++ replace/src/main/kotlin/com/app/replace/ui/response/CommonExceptionFormat.kt
@@ -1,0 +1,3 @@
+package com.app.replace.ui.response
+
+data class CommonExceptionFormat(val errorCode: Int, val errorBody: String)
