+++ replace/src/main/kotlin/com/app/replace/ui/response/ImageUploadingResponse.kt
@@ -1,0 +1,3 @@
+package com.app.replace.ui.response
+
+data class ImageUploadingResponse(val imageUrls: List<String>)
