+++ replace/src/main/kotlin/com/app/replace/ui/response/LoadingSingleConnectionCodeResponse.kt
@@ -1,0 +1,3 @@
+package com.app.replace.ui.response
+
+data class LoadingSingleConnectionCodeResponse(val code : String)
