+++ replace/src/test/kotlin/com/app/replace/application/UserServiceTest.kt
@@ -1,5 +1,7 @@
 package com.app.replace.application
 
+import com.app.replace.application.response.AloneUserInformation
+import com.app.replace.application.response.UserInformationWithPartner
 import com.app.replace.domain.UserRepository
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.matchers.shouldBe
